webapp/requirements 2.txt
acme==1.1.0
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
blinker==1.4
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
connexion==2.7.0
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
email-validator==1.1.0
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-HTTPAuth==4.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
future==0.18.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==16.10.1
inflection==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
mock==3.0.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
openapi-spec-validator==0.2.8
parsedatetime==2.4
passlib==1.7.2
pbr==5.4.5
Pillow==7.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.16.0
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.19
ssh-import-id==5.10
stripe==2.51.0
svgwrite==1.4
systemd-python==234
Tree==0.2.4
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
wadllib==1.3.3
Werkzeug==1.0.1
WTForms==2.3.1
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
